Added error raise for vina docking that doesnt find any poses

+++ reward/docking.py
@@ -151,6 +151,10 @@ def vina_opt(guestmol, vinaobj):
     pdbqt_mol = PDBQTMolecule.from_file(f'reward/vina_pose_{len(curr)+1}.pdbqt', skip_typing=True)
     # Returns a list of rdkit mols
     rdkitmols = RDKitMolCreate.from_pdbqt_mol(pdbqt_mol)
+    
+    # Error in case no poses are found
+    if rdkitmols == []:
+        raise ValueError("No poses found")
 
     os.remove(f'reward/vina_pose_{len(curr)+1}.pdbqt')
